refactor: product 패키지 변경
