+++ src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto.response;
+package kitchenpos.product.dto;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 import java.math.BigDecimal;
 import java.util.List;
