+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,10 +1,9 @@
-package kitchenpos.dao;
+package kitchenpos.product.repository;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
