+++ src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
